all buttons work as intended, still need to add clear button to wipe fields and history

+++ server/server.js
@@ -12,8 +12,8 @@ calcHistory = []
 //app.post below
 app.post('/calc', (req, res) => {
     let calc = req.body;
-    let firstVal = parseInt(calc.valOne);
-    let secVal = parseInt(calc.valTwo);
+    let firstVal = parseFloat(calc.valOne);
+    let secVal = parseFloat(calc.valTwo);
     let operator = calc.operation
     
     switch (operator) {
